+++ src/Modal/index.js
@@ -10,8 +10,10 @@ export default class Modal extends Component {
 
   static propTypes = {
     target: PropTypes.element.isRequired,
+    cancelButtonClass: PropTypes.string,
     cancelLabel: PropTypes.string,
     children: PropTypes.node,
+    confirmButtonClass: PropTypes.string,
     confirmLabel: PropTypes.string,
     isOpen: PropTypes.bool,
     type: PropTypes.string,
@@ -46,12 +48,15 @@ export default class Modal extends Component {
     setTimeout(() => this.setState({ isOpen: true }), 0)
   }
 
-  handleClose = () => {
+  handleClose = (callback) => {
     this.setState(
       { isOpen: false },
       () => setTimeout(
         () => {
           this.modal.closePortal()
+          if (typeof callback === 'function') {
+            callback()
+          }
         },
         300
       )
@@ -106,14 +111,31 @@ export default class Modal extends Component {
 
     let footer = []
     if (type === 'alert') {
-      footer = [() => <Button primary onClick={handleConfirm}>{confirmLabel}</Button>]
+      footer = [
+        () =>
+          <Button
+            primary
+            className={this.props.confirmButtonClass}
+            onClick={handleConfirm}
+          >{confirmLabel}
+          </Button>,
+      ]
     }
 
     if (type === 'confirm' || type === 'prompt') {
       footer = [
-        () => <Button onClick={handleCancel}>{cancelLabel}</Button>,
+        () => <Button className={this.props.cancelButtonClass} onClick={handleCancel}>
+          {cancelLabel}
+        </Button>,
         () => <span>&nbsp;</span>,
-        () => <Button primary onClick={handleConfirm}>{confirmLabel}</Button>,
+        () =>
+          <Button
+            primary
+            className={this.props.confirmButtonClass}
+            onClick={handleConfirm}
+          >
+            {confirmLabel}
+          </Button>,
       ]
     }
     const { children, ...dialogProps } = this.props